Refactored duplicated code
