+++ python-client/giskard/ml_worker/testing/registry/slicing_function.py
@@ -1,16 +1,13 @@
 import functools
 import inspect
-import pickle
-import sys
 from pathlib import Path
 from typing import Optional, List, Union, Type, Callable, Dict
 
-import cloudpickle
 import pandas as pd
 
 from giskard.core.core import DatasetProcessFunctionMeta, DatasetProcessFunctionType
 from giskard.core.validation import configured_validate_arguments
-from giskard.ml_worker.core.savable import Artifact
+from giskard.ml_worker.core.savable import RegistryArtifact
 from giskard.ml_worker.testing.registry.decorators_utils import validate_arg_type, drop_arg, \
     make_all_optional_or_suite_input, set_return_type
 from giskard.ml_worker.testing.registry.registry import get_object_uuid, tests_registry
@@ -20,7 +17,7 @@ SlicingFunctionType = Callable[..., bool]
 default_tags = ['filter']
 
 
-class SlicingFunction(Artifact[DatasetProcessFunctionMeta]):
+class SlicingFunction(RegistryArtifact[DatasetProcessFunctionMeta]):
     """
     A slicing function used to subset data.
 
@@ -93,41 +90,10 @@ class SlicingFunction(Artifact[DatasetProcessFunctionMeta]):
         else:
             return self.func(data, **self.params)
 
-    def _save_locally(self, local_dir: Path):
-        with open(Path(local_dir) / 'data.pkl', 'wb') as f:
-            cloudpickle.dump(self, f, protocol=pickle.DEFAULT_PROTOCOL)
-
-    @classmethod
-    def _load_meta_locally(cls, local_dir, uuid: str) -> Optional[DatasetProcessFunctionMeta]:
-        meta = tests_registry.get_test(uuid)
-
-        if meta is not None:
-            return meta
-
-        return super()._load_meta_locally(local_dir, uuid)
-
     @classmethod
     def load(cls, local_dir: Path, uuid: str, meta: DatasetProcessFunctionMeta):
         if meta.process_type == DatasetProcessFunctionType.CODE:
-            _slicing_function: Optional[SlicingFunction]
-            if local_dir.exists():
-                with open(local_dir / 'data.pkl', 'rb') as f:
-                    _slicing_function = cloudpickle.load(f)
-            else:
-                try:
-                    func = getattr(sys.modules[meta.module], meta.name)
-
-                    if inspect.isclass(func) or hasattr(func, 'meta'):
-                        _slicing_function = func()
-                    else:
-                        _slicing_function = cls(func)
-                        _slicing_function.meta = meta
-                except Exception:
-                    return None
-
-            tests_registry.register(_slicing_function.meta)
-
-            return _slicing_function
+            super().load(local_dir, uuid, meta)
         else:
             return cls._load_no_code(meta)
