+++ python-client/giskard/ml_worker/testing/registry/transformation_function.py
@@ -1,16 +1,12 @@
 import functools
 import inspect
-import pickle
-import sys
-from pathlib import Path
 from typing import Optional, List, Union, Type, Callable
 
-import cloudpickle
 import pandas as pd
 
 from giskard.core.core import DatasetProcessFunctionMeta
 from giskard.core.validation import configured_validate_arguments
-from giskard.ml_worker.core.savable import Artifact
+from giskard.ml_worker.core.savable import RegistryArtifact
 from giskard.ml_worker.testing.registry.decorators_utils import (
     validate_arg_type,
     drop_arg,
@@ -24,7 +20,7 @@ TransformationFunctionType = Callable[..., Union[pd.Series, pd.DataFrame]]
 default_tags = ['transformation']
 
 
-class TransformationFunction(Artifact[DatasetProcessFunctionMeta]):
+class TransformationFunction(RegistryArtifact[DatasetProcessFunctionMeta]):
     func: TransformationFunctionType = None
     row_level: bool = True
     cell_level: bool = False
@@ -70,44 +66,6 @@ class TransformationFunction(Artifact[DatasetProcessFunctionMeta]):
             return data.apply(lambda row: self.func(row, **self.params), axis=1)
         else:
             return self.func(data, **self.params)
-
-    def _should_upload(self) -> bool:
-        return self.meta.version is None
-
-    def _save_locally(self, local_dir: Path):
-        with open(Path(local_dir) / 'data.pkl', 'wb') as f:
-            cloudpickle.dump(self, f, protocol=pickle.DEFAULT_PROTOCOL)
-
-    @classmethod
-    def _load_meta_locally(cls, local_dir, uuid: str) -> Optional[DatasetProcessFunctionMeta]:
-        meta = tests_registry.get_test(uuid)
-
-        if meta is not None:
-            return meta
-
-        return super()._load_meta_locally(local_dir, uuid)
-
-    @classmethod
-    def load(cls, local_dir: Path, uuid: str, meta: DatasetProcessFunctionMeta):
-        _transformation_function: Optional[TransformationFunction]
-        if local_dir.exists():
-            with open(local_dir / 'data.pkl', 'rb') as f:
-                _transformation_function = cloudpickle.load(f)
-        else:
-            try:
-                func = getattr(sys.modules[meta.module], meta.name)
-
-                if inspect.isclass(func) or hasattr(func, 'meta'):
-                    _transformation_function = func()
-                else:
-                    _transformation_function = cls(func)
-                    _transformation_function.meta = meta
-            except Exception:
-                return None
-
-        tests_registry.register(_transformation_function.meta)
-
-        return _transformation_function
 
     @classmethod
     def _get_meta_class(cls):
